+++ Rhygine/UIRootElement.cpp
@@ -14,7 +14,7 @@ void UIRootElement::Delete()
 
 void UIRootElement::InnerDelete(UIElement* element)
 {
-	for (int i = element->children.size() - 1; i > -1; i--)
+	for (int i = static_cast<int>(element->children.size()) - 1; i > -1; i--)
 		InnerDelete(element->children[i]);
 	element->GetGameObject()->RemoveComponent(element);
 }
