+++ Rhygine/UIRootSizer.cpp
@@ -8,8 +8,8 @@ const bool UIRootSizer::isControllingChildren()
 
 void UIRootSizer::ResizeSelf(RhyM::Rect& rect, RhyM::Vec2 currentWorldScale)
 {
-	rect.width = Window::GetInstance()->GetWidth();
-	rect.height = Window::GetInstance()->GetHeight();
+	rect.width = static_cast<float>(Window::GetInstance()->GetWidth());
+	rect.height = static_cast<float>(Window::GetInstance()->GetHeight());
 }
 
 void UIRootSizer::UpdatePositionSelf(RhyM::Rect& rect)
